Refactor: Move queue specific methods to queues.c
